prepare barcode for input+camera scanner options
