+++ frontend/react-app/src/components/barcode-scanner/BarCodeScannerModal.tsx
@@ -1,40 +1,21 @@
-import { useRef } from "react";
 import styled from "styled-components";
-import Scanner from "./Scanner";
-import { useFetchCoordsByBarcode } from "./useFetchCoordsByBarcode";
-import { debounce } from "../../utils";
-import { Loader } from "../core/Loader";
 import { Modal } from "../core/Modal";
+import { CameraScanner } from "./CameraScanner";
 
-const ResultText = styled.p`
-  font-size: 1.5rem;
-  color: white;
-`;
+export const BarcodeScannerModal = ({ toggleModal }: any) => {
+  // TODO set state for "option"
 
-export const BarcodeScannerModal = ({ isModalOpen, toggleModal }: any) => {
-  const { error, isLoading, isSuccess, fetchCoordsByBarcode } =
-    useFetchCoordsByBarcode();
-  const barcodeScanned = useRef<string>(null);
+  // TODO: endre denne modalen
+  // lag state for option ("manuell", "kamera")
+  // vise to knapper:
+  // [legg inn strekkkode manuelt] --> onclick --> rendrer kun inputfelt med knapp "scan"
+  // [bruk kamera] --> onclick --> render <CameraScanner>
 
-  const onDetected = debounce((barcode: string) => {
-    if (barcode !== barcodeScanned.current) {
-      barcodeScanned.current = barcode;
-      fetchCoordsByBarcode(barcode);
-    }
-  }, 500);
-
-  if (!isModalOpen) return null;
-  if (isSuccess) toggleModal();
+  // rendringen skal skje under knappene, så brukeren kan velge mellom kamera/inputfelt frem og tilbake
 
   return (
     <Modal onClose={toggleModal}>
-      {isLoading && (
-        <>
-          <ResultText>Henter avfallsdata</ResultText> <Loader />
-        </>
-      )}
-      {error && <ResultText>{error}</ResultText>}
-      <Scanner onDetected={onDetected} />
+      <CameraScanner toggleModal={toggleModal} />
     </Modal>
   );
 };
